requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Database Connectivity
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Azure Integration
azure-storage-blob>=12.0.0
azure-identity>=1.7.0
azure-mgmt-storage>=18.0.0

# Configuration and Utilities
pyyaml>=6.0
python-dotenv>=0.19.0

# Data Processing
openpyxl>=3.0.0
xlrd>=2.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Logging and Monitoring
structlog>=22.0.0

# Development Tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991 